+Added
-- Byron -- Added initial files to the github -- Please always use
dashcode when doing this project, for us to be consistent -- To properly
utilize our work environment, sign up for github, and download github
for mac found in their site. -- Second, find a way to add this to your
list of repositories https://github.com/butchthecat/tipid -- If you have
question, PM me.